# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Skymark Eatery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, confirmed, preparing, ready, completed, cancelled, null]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /payments/create-intent:
    post:
      operationId: createPaymentIntent
      tags: [payments]
      summary: Create a Stripe payment intent for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentBody"
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResponse"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment and finalize order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        specialInstructions:
          type: ["string", "null"]
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice, subtotal]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        total:
          type: number
        stripePaymentIntentId:
          type: ["string", "null"]
        stripeStatus:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        estimatedReadyAt:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, customerEmail, status, total, items, createdAt, updatedAt]
    CreateOrderItemBody:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        specialInstructions:
          type: ["string", "null"]
      required: [menuItemId, quantity]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required: [customerName, customerEmail, items]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        estimatedReadyAt:
          type: ["string", "null"]
      required: [status]

    # ─── PAYMENTS ──────────────────────────────────────────────────────────────
    CreatePaymentIntentBody:
      type: object
      properties:
        orderId:
          type: integer
      required: [orderId]
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: integer
      required: [clientSecret, paymentIntentId, amount]
    ConfirmPaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        paymentIntentId:
          type: string
      required: [orderId, paymentIntentId]

    # ─── ADMIN ─────────────────────────────────────────────────────────────────
...