# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  items: CreateOrderItemBody[];
}