# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  stripeStatus?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  /** @nullable */
  estimatedReadyAt?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}